Refactor alias tool imports and introduce main functionality
- Updated import statements in the experimental tools package to source the main alias functionality from the new `main.py` module.
- Added a new `main.py` file in the alias tool, implementing the core command-line interface for managing URL aliases, including listing and replacing aliases.
- Enhanced argument parsing and logging setup for improved usability and traceability in alias operations.
