+++ src/tlc_tools/experimental/__init__.py
@@ -1,4 +1,4 @@
 """Experimental tools package."""
 
-from .alias_tool.alias import *  # noqa: F401, F403
+from .alias_tool.main import *  # noqa: F401, F403
 from .augment_tool.augment_bb_table import *  # noqa: F401, F403
